backend/requirements.txt
fastapi[standard]
uvicorn
google-genai
python-multipart
python-dotenv
